requirements.txt
paramiko
pillow
pyautogui
